Fix conn close issues using a specific ctx
Managing the connection lifecycle using a channel (conn.closing) that is
separate from the connection's context may be lead to race condition
errors. Instead, this change creates a new context for the connection
and cancels it when the browser is closing. Since the context is
connected to the parent ones, it can automatically put the connection in
a shutdown/closing stage even if the browser.close is not called, but
happens because one of the parent contextes are canceled.