Do not build browser context based on VU context
Because we want to handle browser lifecycle automatically from our Go
implementation, we can not build the browser context based on the VU
context, which will be closed at the end of every iteration. That would
result in a race condition between iteration and browser closing
procedure. Instead we will wait for the VU context to be closed and then
from our implementation explicitely call Browser.Close() method.