Initialize browser on newPage
The browser.newPage() method will automatically initialize or retrieve a
previously initialized browser for the scenario and iteration. This
browser is held at the root module level pool throughout the iteration
lifecycle. When the iteration ends, signaled by the VU context being
done, the browser is closed and removed from the pool.
If the browser for the scenario and iteration does not exist in the
pool, a new browser process is initialized based on the browser process
for the handle (local or remote CDP WS URL) stored at the browser
process pool. If there is not a browser process initialized for the
handle, a new browser process is launched/connected to and stored in the
pool.