Cancel browser context on close
Previously, because browser context was a child from VU context, it was
closed at the end of every iteration. Now, because browser context is
inherited from background context, we have to cancel it explicitely in
order to notify goroutines that depend on it, such as the browser and
frame session events goroutines. In many cases these goroutines were
already exiting successfully because the browser context was being
canceled in reaction to the EventConnectionClose received by the
browser events reading goroutine (see:
https://github.com/grafana/xk6-browser/blob/4193ba72dd57d711ac27cab6781502a03ec2132f/common/browser.go#L199-L201).
But still it's better to also cancel the context explicitly from within
the browser closing procedure.