Fix test browser cancel
Now that browser does not depend on the VU context, we have to export
through the browser type implementation the browser context cancel
function so browser can be canceled from test browser for testing
purposes. Additionally, because the browser context does no longer
depend on the VU context but on the context created during browser type
init, we have to skip close on cleanup if bt.Ctx is done, instead of
depending on the VU context.